+++ test/conftest.py
@@ -1,22 +1,41 @@
-from test.fixtures import init_acts, init_methods
-from types import SimpleNamespace
+from dataclasses import dataclass
+
+from bw2data.backends import Activity
+
+from test.fixtures import init_methods
 
 import bw2data
-import bw2io
 import pytest
 
 from lca_algebraic import newActivity, resetDb, resetParams
 from lca_algebraic.cache import clear_caches
-from lca_algebraic.settings import Settings
 
 USER_DB = "fg"
 BG_DB = "bg"
 METHOD_PREFIX = "tests"
 TST_PROJECT_NAME = "tests-bw25"
 
+MethodKey = tuple[str, str, str]
+
+
+@dataclass
+class DataFixture:
+    bio1: Activity
+    bio2: Activity
+    bio3: Activity
+
+    bg_act1: Activity
+    bg_act2: Activity
+    bg_act3: Activity
+
+    ibio1: MethodKey
+    ibio2: MethodKey
+    ibio3: MethodKey
+    imulti: MethodKey
+
 
 @pytest.fixture(autouse=True, scope="module")
-def data():
+def data() -> DataFixture:
     """Setup background data"""
 
     # Reset func project, empty DB
@@ -45,7 +64,7 @@ def data():
     # Cleanup foreground DB
     resetDb(USER_DB, True)
 
-    return SimpleNamespace(**locals())
+    return DataFixture(bio1, bio2, bio3, bg_act1, bg_act2, bg_act3, ibio1, ibio2, ibio3, imulti)
 
 
 @pytest.fixture(autouse=True, scope="function")